+++ ve-silo/contracts/gauges/CHANGELOG.md
@@ -1,5 +1,7 @@
 ### ethereum/SiloLiquidityGauge.vy
 
+- Fees receivers from the Silo factory and fees for the gauge from the MainnetBalancerMinter.sol ([35c6ac8](https://github.com/silo-finance/silo-contracts-v2/pull/155/commits/35c6ac87bf4a327758fa22486ef7cf2aec7810f0))
+
 - Implemented `claimable_tokens_with_fees` fn, which returns claimable tokens for a user and also fees gained by a DAO and deployer ([a96d717](https://github.com/silo-finance/silo-contracts-v2/pull/97/commits/a96d7173319a766536acc4874aa9b1670c54f0fa))
 
 - Replaced an ERC-20 Balancer handler with a Silo Hook Receiver. Added integration with the Silo share token ([ccf97d1](https://github.com/silo-finance/silo-contracts-v2/pull/94/commits/ccf97d1d8434ac54c169e954094a3d4537ed5b3d))
@@ -35,6 +37,8 @@
 - Copy of Balancer's implementation of the LiquidityGaugeFactory.sol ([c2bc3d5](https://github.com/silo-finance/silo-contracts-v2/pull/34/commits/c2bc3d539244abee8e2cd9b13e70b931eb251735))
 
 ### l2-common/ChildChainGauge.vy
+
+- Fees receivers from the Silo factory and fees for the gauge from the L2BalancerPseudoMinter.sol ([3f6faaa](https://github.com/silo-finance/silo-contracts-v2/pull/155/commits/3f6faaafe909da7aa2a064660f02c8233e4be86a))
 
 - Implemented `claimable_tokens_with_fees` fn, which returns claimable tokens for a user and also fees gained by a DAO and deployer ([5611fe5](https://github.com/silo-finance/silo-contracts-v2/pull/97/commits/5611fe5eb81553e6a47db71638a601346fad065e))